Select Service Assignment
From here on in the Assignments will use the ShowTracker database. We
will also use ADO Data Entities rather than classic AD0.

In this assignment we will create a new service that returns various
data from the database. In particular it should
list of all venues (just the names of the venues)
A list of all artists (just the names)
list of all shows (just the names)
list of all shows for a particular venue (for this list the show name,
show date,  show start time)
list of all shows for a particular artist (Show name, show date, show
time and venue Name)
Additionally Create tests for each of the methods and test whether they
succeed or fail.